Модифицирована функция. Принимает вторым параметром Boolean.

+++ app/src/main/java/com/ggr3ml1n/shoppinglist/vm/MainViewModel.kt
@@ -47,8 +47,8 @@ class MainViewModel(database: MainDataBase) : ViewModel() {
         dao.deleteNote(id)
     }
 
-    fun deleteShopList(id: Int) = viewModelScope.launch {
-        dao.deleteShopListName(id)
+    fun deleteShopList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+        if (deleteList) dao.deleteShopListName(id)
         dao.deleteShopListItemsByListId(id)
     }
